+++ extensions/iceworks-app/src/extension.ts
@@ -9,6 +9,7 @@ import { createComponentsTreeProvider } from './views/componentsView';
 import { createPagesTreeProvider } from './views/pagesView';
 import { ITerminalMap } from './types';
 import services from './services';
+import { createStatusBarItem, openCommandPaletteCommandId, registerOpenCommandPalette } from './createStatusBarItem';
 
 // eslint-disable-next-line
 const { name, version } = require('../package.json');
@@ -28,6 +29,10 @@ export async function activate(context: vscode.ExtensionContext) {
       version,
     }
   });
+  // init statusBarItem
+  const statusBarItem = createStatusBarItem();
+  subscriptions.push(vscode.commands.registerCommand(openCommandPaletteCommandId, registerOpenCommandPalette));
+  subscriptions.push(statusBarItem);
   // init webview
   function activeWebview() {
     const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', 'Iceworks 设置', ViewColumn.One, {
@@ -45,13 +50,18 @@ export async function activate(context: vscode.ExtensionContext) {
   if (!rootPath) {
     vscode.window.showInformationMessage('当前工作区为空，请打开应用或新建应用。');
     vscode.commands.executeCommand('setContext', 'iceworks:isNotTargetProject', true);
+    vscode.commands.executeCommand('iceworks-project-creator.start');
     return;
   }
   try {
     const projectType = await getProjectType();
-    vscode.commands.executeCommand('setContext', 'iceworks:isNotTargetProject', projectType === 'unknown');
+    const isNotTargetProject = projectType === 'unknown';
+    vscode.commands.executeCommand('setContext', 'iceworks:isNotTargetProject', isNotTargetProject);
+    if (isNotTargetProject)
+      vscode.commands.executeCommand('iceworks-project-creator.start');
   } catch (e) {
     vscode.commands.executeCommand('setContext', 'iceworks:isNotTargetProject', true);
+    vscode.commands.executeCommand('iceworks-project-creator.start');
   }
   const terminals: ITerminalMap = new Map<string, Terminal>();
   // init tree data providers