+++ extensions/iceworks-app/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.34
 
+- feat: add Iceworks status bar entry
+- feat: start the iceworks-project-creator automatically
 - docs: Project => Application
 
 ## 0.1.33